for zabbix
this script aims to work exclusively under Linux OS. The serial port is rigidly fixed in the script for specific equipment that will transmit data to the server through it. Also now the information that comes to the serial port is parsed and formed to be sent to the ZABBIX file, from which the program understands from what equipment and from what port errors come. The script is run in a perpetual loop, as well as the program implements a constant poll of the serial port for its status. in case of connection failure, the program does not close, but continues to poll the port until the connection appears and sends messages about the failure to ZABBIX